refactor(ci): move the Windows subprocess cap into a build-tagged file
Splits the semaphore that caps concurrent subprocess launches on Windows
(added in 66e50d45f5, scoped to Windows in 72319350ae) out of command.go's
shared runtime.GOOS check into subprocess_cap_windows.go / subprocess_cap_other.go,
matching this codebase's established *_windows.go/*_other.go build-tag
convention (e.g. pkg/terraform/cache/trust_install_windows.go) instead of
a runtime branch in shared code.

Also adds the docs/fixes/ record for the whole fix (both prior commits plus
this refactor), which was missing.
