+++ internal/ci/acceptance/command.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 )
@@ -35,48 +34,7 @@ const (
 	// can't grow that buffer without limit. The real diagnostic is one short line
 	// (well under this), so bounding it doesn't affect detection.
 	maxTransientMatchWindow = 4096
-
-	// Caps how many subprocesses (go build/go test/go tool covdata, plus each
-	// precompiled *.test.exe) this package launches at once, on Windows hosts
-	// only (see acquireSubprocessSlot). This package's own acceptance tests run
-	// dozens of t.Parallel() subtests that each shell out, so without a cap, a
-	// wide/many-core Windows CI runner lets that many real `go` toolchain
-	// invocations (each independently allocating and syscalling heavily) run
-	// fully concurrently. That has produced a runtime-fatal GC/allocator crash
-	// ("fatal error: found pointer to free object" / "marked free object in
-	// span") in mcache/mgcsweep during a concurrent os/exec process launch -- a
-	// long-standing, still-recurring class of Go runtime race under heavy
-	// concurrent allocation+syscall pressure specifically on Windows (see e.g.
-	// golang/go#44900, #45364, #47415, #54247; the reports are Windows-only),
-	// not anything specific to the command being run. Serializing actual
-	// subprocess launches (while still letting the surrounding Go test logic
-	// run in parallel) avoids the trigger condition without giving up test
-	// parallelism where it doesn't involve a real subprocess -- and without
-	// slowing down Linux/macOS, which haven't exhibited this crash.
-	maxConcurrentSubprocesses = 4
 )
-
-// subprocessSlots limits how many commandRunner.run/output calls -- across every
-// commandRunner instance, since each caller constructs its own -- may have a real
-// `cmd.Run()` in flight at once, on Windows. See maxConcurrentSubprocesses.
-var subprocessSlots = make(chan struct{}, maxConcurrentSubprocesses)
-
-// acquireSubprocessSlot blocks until a subprocess slot is free or ctx is done,
-// returning a release func to call (typically deferred) once the subprocess
-// exits. It's a no-op on every host OS but Windows -- the GC/allocator race
-// this guards against (see maxConcurrentSubprocesses) has only been observed
-// there, so Linux/macOS runs keep their full, uncapped concurrency.
-func acquireSubprocessSlot(ctx context.Context) (func(), error) {
-	if runtime.GOOS != "windows" {
-		return func() {}, nil
-	}
-	select {
-	case subprocessSlots <- struct{}{}:
-		return func() { <-subprocessSlots }, nil
-	case <-ctx.Done():
-		return nil, ctx.Err()
-	}
-}
 
 var (
 	errInvalidConfiguration = errors.New("invalid acceptance configuration")
